refactored get_bucket_encryption
