+++ examples/file_downloader.rs
@@ -1,27 +1,30 @@
-use minio::s3::args::{BucketExistsArgs, MakeBucketArgs};
+// MinIO Rust Library for Amazon S3 Compatible Cloud Storage
+// Copyright 2025 MinIO, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+mod common;
+
+use crate::common::{create_bucket_if_not_exists, create_client_on_play};
 use minio::s3::builders::ObjectContent;
-use minio::s3::client::ClientBuilder;
-use minio::s3::creds::StaticProvider;
-use minio::s3::http::BaseUrl;
 use minio::s3::types::S3Api;
+use minio::s3::Client;
 use std::path::Path;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     env_logger::init(); // Note: set environment variable RUST_LOG="INFO" to log info and higher
-
-    let base_url = "https://play.min.io".parse::<BaseUrl>()?;
-    log::info!("Trying to connect to MinIO at: `{:?}`", base_url);
-
-    let static_provider = StaticProvider::new(
-        "Q3AM3UQ867SPQQA43P2F",
-        "zuf+tfteSlswRu7BJ86wekitnifILbZam1KYY3TG",
-        None,
-    );
-
-    let client = ClientBuilder::new(base_url.clone())
-        .provider(Some(Box::new(static_provider)))
-        .build()?;
+    let client: Client = create_client_on_play()?;
 
     let bucket_name: &str = "file-download-rust-bucket";
     let object_name: &str = "cat.png";
@@ -30,19 +33,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     let filename: &Path = Path::new("./examples/cat.png");
     let download_path: &str = &format!("/tmp/downloads/{object_name}");
 
-    // Check 'bucket_name' bucket exist or not.
-    let exists: bool = client
-        .bucket_exists(&BucketExistsArgs::new(bucket_name).unwrap())
-        .await
-        .unwrap();
-
-    // Make 'bucket_name' bucket if not exist.
-    if !exists {
-        client
-            .make_bucket(&MakeBucketArgs::new(bucket_name).unwrap())
-            .await
-            .unwrap();
-    }
+    create_bucket_if_not_exists(bucket_name, &client).await?;
 
     if filename.exists() {
         log::info!("File '{}' exists.", &filename.to_str().unwrap());
@@ -64,10 +55,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
 
     let get_object = client.get_object(bucket_name, object_name).send().await?;
 
-    get_object
-        .content
-        .to_file(&Path::new(download_path))
-        .await?;
+    get_object.content.to_file(Path::new(download_path)).await?;
 
     log::info!("Object '{object_name}' is successfully downloaded to file '{download_path}'.");
