+++ examples/object_prompt.rs
@@ -13,7 +13,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use minio::s3::args::{BucketExistsArgs, MakeBucketArgs};
+mod common;
+
+use crate::common::create_bucket_if_not_exists;
 use minio::s3::builders::{ObjectContent, ObjectPrompt};
 use minio::s3::client::ClientBuilder;
 use minio::s3::creds::StaticProvider;
@@ -25,14 +27,11 @@ use std::path::Path;
 async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     env_logger::init(); // Note: set environment variable RUST_LOG="INFO" to log info and higher
 
-    let base_url = "https://play.min.io".parse::<BaseUrl>()?;
+    //Note: object prompt is not supported on play.min.io, you will need point to AIStor
+    let base_url = "http://localhost:9000".parse::<BaseUrl>()?;
     log::info!("Trying to connect to MinIO at: `{:?}`", base_url);
 
-    let static_provider = StaticProvider::new(
-        "Q3AM3UQ867SPQQA43P2F",
-        "zuf+tfteSlswRu7BJ86wekitnifILbZam1KYY3TG",
-        None,
-    );
+    let static_provider = StaticProvider::new("admin", "admin", None);
 
     let client = ClientBuilder::new(base_url.clone())
         .provider(Some(Box::new(static_provider)))
@@ -40,20 +39,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
         .build()?;
 
     let bucket_name: &str = "object-prompt-rust-bucket";
-
-    // Check 'bucket_name' bucket exist or not.
-    let exists: bool = client
-        .bucket_exists(&BucketExistsArgs::new(bucket_name).unwrap())
-        .await
-        .unwrap();
-
-    // Make 'bucket_name' bucket if not exist.
-    if !exists {
-        client
-            .make_bucket(&MakeBucketArgs::new(bucket_name).unwrap())
-            .await
-            .unwrap();
-    }
+    create_bucket_if_not_exists(bucket_name, &client).await?;
 
     // File we are going to upload to the bucket
     let filename: &Path = Path::new("./examples/cat.png");
