+++ src/s3/builders/get_bucket_versioning.rs
@@ -1,0 +1,56 @@
+// MinIO Rust Library for Amazon S3 Compatible Cloud Storage
+// Copyright 2025 MinIO, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::s3::builders::BucketCommon;
+use crate::s3::error::Error;
+use crate::s3::response::GetBucketVersioningResponse;
+use crate::s3::types::{S3Api, S3Request, ToS3Request};
+use crate::s3::utils::{merge, Multimap};
+use http::Method;
+
+/// Argument builder for [get_bucket_versioning()](Client::get_bucket_versioning) API
+pub type GetBucketVersioning = BucketCommon<GetBucketVersioningPhantomData>;
+
+#[derive(Default, Debug)]
+pub struct GetBucketVersioningPhantomData;
+
+impl S3Api for GetBucketVersioning {
+    type S3Response = GetBucketVersioningResponse;
+}
+
+impl ToS3Request for GetBucketVersioning {
+    fn to_s3request(&self) -> Result<S3Request, Error> {
+        let mut headers = Multimap::new();
+        if let Some(v) = &self.extra_headers {
+            merge(&mut headers, v);
+        }
+
+        let mut query_params = Multimap::new();
+        if let Some(v) = &self.extra_query_params {
+            merge(&mut query_params, v);
+        }
+        query_params.insert(String::from("versioning"), String::new());
+
+        let req = S3Request::new(
+            self.client.as_ref().ok_or(Error::NoClientProvided)?,
+            Method::GET,
+        )
+        .region(self.region.as_deref())
+        .bucket(Some(&self.bucket))
+        .query_params(query_params)
+        .headers(headers);
+        Ok(req)
+    }
+}
