+++ src/s3/builders/object_content.rs
@@ -200,10 +200,11 @@ impl ObjectContent {
             .to_path_buf()
             .join(Path::new(tmp_file_name.as_os_str()));
 
-        let mut total = 0;
+        let mut total_bytes_written = 0;
         let mut fp = fs::OpenOptions::new()
             .write(true)
-            .create(true)
+            .create(true) // Ensures that the file will be created if it does not already exist
+            .truncate(true) // Clears the contents (truncates the file size to 0) before writing
             .open(&tmp_file_path)
             .await?;
         let (mut r, _) = self.to_stream().await?;
@@ -212,12 +213,12 @@ impl ObjectContent {
             if bytes.is_empty() {
                 break;
             }
-            total += bytes.len() as u64;
+            total_bytes_written += bytes.len() as u64;
             fp.write_all(&bytes).await?;
         }
         fp.flush().await?;
         fs::rename(&tmp_file_path, file_path).await?;
-        Ok(total)
+        Ok(total_bytes_written)
     }
 }
