+++ src/s3/response.rs
@@ -25,22 +25,26 @@ use xmltree::Element;
 use crate::s3::error::Error;
 use crate::s3::types::{
     parse_legal_hold, LifecycleConfig, NotificationConfig, ObjectLockConfig, ReplicationConfig,
-    RetentionMode, SelectProgress, SseConfig,
+    RetentionMode, SelectProgress,
 };
 use crate::s3::utils::{
     copy_slice, crc32, from_http_header_value, from_iso8601utc, get_text, uint32, UtcTime,
 };
 
-mod buckets;
+mod get_bucket_encryption;
+mod get_bucket_versioning;
 mod get_object;
+mod list_buckets;
 pub(crate) mod list_objects;
 mod listen_bucket_notification;
 mod object_prompt;
 mod put_object;
 mod remove_objects;
 
-pub use buckets::{GetBucketVersioningResponse, ListBucketsResponse};
+pub use get_bucket_encryption::GetBucketEncryptionResponse;
+pub use get_bucket_versioning::GetBucketVersioningResponse;
 pub use get_object::GetObjectResponse;
+pub use list_buckets::ListBucketsResponse;
 pub use list_objects::ListObjectsResponse;
 pub use listen_bucket_notification::ListenBucketNotificationResponse;
 pub use object_prompt::ObjectPromptResponse;
@@ -549,15 +553,6 @@ impl SelectObjectContentResponse {
 
 /// Response of [delete_bucket_encryption()](crate::s3::client::Client::delete_bucket_encryption) API
 pub type DeleteBucketEncryptionResponse = BucketResponse;
-
-#[derive(Clone, Debug)]
-/// Response of [get_bucket_encryption()](crate::s3::client::Client::get_bucket_encryption) API
-pub struct GetBucketEncryptionResponse {
-    pub headers: HeaderMap,
-    pub region: String,
-    pub bucket_name: String,
-    pub config: SseConfig,
-}
 
 /// Response of [set_bucket_encryption()](crate::s3::client::Client::set_bucket_encryption) API
 pub type SetBucketEncryptionResponse = BucketResponse;
