+++ tests/tests.rs
@@ -39,13 +39,12 @@ use minio::s3::client::Client;
 use minio::s3::creds::StaticProvider;
 use minio::s3::error::Error;
 use minio::s3::http::BaseUrl;
-use minio::s3::types::ToStream;
+use minio::s3::response::GetBucketVersioningResponse;
 use minio::s3::types::{
     CsvInputSerialization, CsvOutputSerialization, FileHeaderInfo, NotificationConfig,
-    ObjectLockConfig, PrefixFilterRule, QueueConfig, QuoteFields, RetentionMode, SelectRequest,
-    SuffixFilterRule,
+    NotificationRecords, ObjectLockConfig, PrefixFilterRule, QueueConfig, QuoteFields,
+    RetentionMode, S3Api, SelectRequest, SuffixFilterRule, ToStream,
 };
-use minio::s3::types::{NotificationRecords, S3Api};
 use minio::s3::utils::{to_iso8601utc, utc_now};
 
 struct RandReader {
@@ -1263,9 +1262,10 @@ impl ClientTest {
             .await
             .unwrap();
 
-        let resp = self
+        let resp: GetBucketVersioningResponse = self
             .client
             .get_bucket_versioning(&bucket_name)
+            .unwrap()
             .send()
             .await
             .unwrap();
@@ -1276,9 +1276,10 @@ impl ClientTest {
             .await
             .unwrap();
 
-        let resp = self
+        let resp: GetBucketVersioningResponse = self
             .client
             .get_bucket_versioning(&bucket_name)
+            .unwrap()
             .send()
             .await
             .unwrap();
